requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.8
botocore==1.38.8
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
distro==1.9.0
entrypoints==0.4
Faker==37.1.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
htbuilder==0.6.2
httpcore==1.0.9
httpx==0.28.1
hydralit_components==1.0.10
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.7.0
narwhals==1.38.0
numpy==2.2.5
openai==1.77.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.7
plotly==5.23.0
prometheus_client==0.20.0
protobuf==5.27.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.15
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
s3transfer==0.12.0
setuptools==80.3.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.31.0
sortedcontainers==2.4.0
soupsieve==2.7
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.37.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-option-menu==0.4.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.33.0
watchdog==4.0.2
wheel==0.45.1
